test: Move insights pruning schedule tests to unit tests
The fake timers seem to affect mysql's connection pooling and causes the
tests to fail on when running with pool size of 1. These tests don't need
to be integration tests to begin with.
